Int+accumlator attacker damage tracking, spawn reset accumlator, fix menu timeout (#497)
* Instead of single float health variable, it's now accumlated quite
  like how `CBaseCombatCharacter::OnTakeDamage_Alive` does it
* Minor refactor to utilize the AttackersTotals struct more
* Fix accumlator not being reset in spawn
* Fix menu timeout being restricted to 5s max, just removed the usage of
  the global timeout
* Attacker info menu timeout increased to 60s
* fixes #309

+++ mp/src/game/shared/neo/neo_gamerules.cpp
@@ -1686,8 +1686,8 @@ static CNEO_Player* FetchAssists(CNEO_Player* attacker, CNEO_Player* victim)
 			continue;
 		}
 
-		const float assistDmg = victim->GetAttackersScores(assistIdx);
-		static const float MIN_DMG_QUALIFY_ASSIST = 50.0f;
+		const int assistDmg = victim->GetAttackersScores(assistIdx);
+		static const int MIN_DMG_QUALIFY_ASSIST = 50;
 		if (assistDmg >= MIN_DMG_QUALIFY_ASSIST)
 		{
 			return static_cast<CNEO_Player*>(UTIL_PlayerByIndex(assistIdx));
